Address IndexedDB migration review feedback
