+++ .changeset/fuzzy-databases-abort.md
@@ -1,0 +1,5 @@
+---
+"@effect/platform-browser": patch
+---
+
+Abort IndexedDB versionchange transactions when schema migrations fail.
